Organize the classes into java packages
