+++ src/main/java/duke/command/MarkCommand.java
@@ -1,19 +1,28 @@
-public class UnmarkCommand extends Command {
+package duke.command;
+
+import duke.task.Task;
+import duke.util.DukeException;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
+public class MarkCommand extends Command {
     private final int taskNum;
 
-    public UnmarkCommand(int taskNum) {
+    public MarkCommand(int taskNum) {
         this.taskNum = taskNum;
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         int currentTotalTasks = taskList.getTotalTasks();
         if (taskNum <= currentTotalTasks && taskNum > 0) {
-            Task taskUnmarked = taskList.unmark(this.taskNum);
-            ui.printOutput("Okay, I've marked this task as undone:\n      " + taskUnmarked);
+            Task taskMarked = taskList.mark(this.taskNum);
+            ui.printOutput("Nice! You've completed this task:\n      " + taskMarked);
             storage.saveToHardDisk(taskList);
         } else {
             throw new DukeException("Invalid task number. There are " + currentTotalTasks + " in the list.");
         }
+
     }
 
     public boolean isExit() {
