+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,20 +1,27 @@
-public class MarkCommand extends Command {
+package duke.command;
+
+import duke.task.Task;
+import duke.util.DukeException;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
+public class UnmarkCommand extends Command {
     private final int taskNum;
 
-    public MarkCommand(int taskNum) {
+    public UnmarkCommand(int taskNum) {
         this.taskNum = taskNum;
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         int currentTotalTasks = taskList.getTotalTasks();
         if (taskNum <= currentTotalTasks && taskNum > 0) {
-            Task taskMarked = taskList.mark(this.taskNum);
-            ui.printOutput("Nice! You've completed this task:\n      " + taskMarked);
+            Task taskUnmarked = taskList.unmark(this.taskNum);
+            ui.printOutput("Okay, I've marked this task as undone:\n      " + taskUnmarked);
             storage.saveToHardDisk(taskList);
         } else {
             throw new DukeException("Invalid task number. There are " + currentTotalTasks + " in the list.");
         }
-
     }
 
     public boolean isExit() {
