+++ src/main/java/duke/util/DukeException.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 public class DukeException extends Exception {
 
     public DukeException(String message) {
