+++ src/main/java/duke/util/Ui.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
